// src/app/modules/academicSemester/academicSemester.interface.ts
import { Model } from 'mongoose';
export type IAcademicSemesterMonths =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December';


  export type IAcademicSemesterTitles = 'Autumn' | 'Summer' | 'Fall';

  export type IAcademicSemesterCode = '01' | '02' | '03';

export type IAcademicSemester = {
  title: IAcademicSemesterTitles
  year: number;
  code: IAcademicSemesterCode
  startMonth: IAcademicSemesterMonths;
  endMonth: IAcademicSemesterMonths;
};

export type AcademicSemeterModel = Model<IAcademicSemester>;